requirements.txt
azureml-core~=1.38.0
azureml-dataprep~=2.29.0
scikit-learn~=0.24.2
pandas~=1.3.3
numpy~=1.21.2